SubNG/requirements.txt
fastapi==0.111.0
telethon==1.35.0
python-dotenv==1.0.1
uvicorn[standard]
